restructure and simplify
