+++ src/c2rust/CMakeLists.txt
@@ -8,13 +8,10 @@ add_library(c2rust STATIC)
 # For the target sources we use a folder `c2rust` such that the main cmake file does not 
 # compile duplicate symbols by its glob * usage.
 target_sources(c2rust PRIVATE 
-    "module-init.c" 
-    "config.c"
-    "value.c"
-    "../util/c2rust/mempool.c"
+    "../module-init/module-init.c" 
+    "../config.c"
+    "../value.c"
+    "../util/mempool.c"
+    #"variadic_helper.c"
 )
 target_include_directories(c2rust PRIVATE . ..)
-
-add_library(c2rust_variadic_helper STATIC)
-target_sources(c2rust_variadic_helper PRIVATE "variadic_helper.c")
-target_include_directories(c2rust_variadic_helper PRIVATE . ..)
