+++ src/module-init/module-init.c
